// src/types/database.ts
export interface EstatusEquipo {
  idEstatus: number;
  estatus: string;
}

export interface TipoEquipo {
  idTipoEquipo: number;
  nombreTipo: string;
  descripcion: string;
}
// ...
export interface VistaEquipoCompleto {
  no_serie: string;
  nombreEquipo: string;
  modelo: string;
  TipoEquipo: string;
  DescripcionTipo: string;
  numeroActivo: string;
  fechaAlta: Date;
  UsuarioAsignado: string;
  EstatusEquipo: string;
  SucursalActual: string;
  AreaActual: string;
  ZonaSucursal: string;
  EstadoSucursal: string;
  MunicipioSucursal: string;
}
// ...
export interface ApiResponse<T> {
  success: boolean;
  data?: T;
  message?: string;
  error?: string;
}